requirements.txt
crewai==0.152.0
#openai==1.98.0
pydantic==2.11.7
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
python-dateutil==2.8.2
urllib3==1.26.18
onnxruntime==1.22.0
google-genai
crewai-tools